pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cluster-triage-env"
version = "1.0.0"
description = "An OpenEnv-compliant RL environment simulating a 4-node enterprise data cluster."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "requests>=2.31.0",
    "pydantic==2.10.4",
    "openai==1.58.1",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",
    "openenv-core<=0.2.0"
]

[project.scripts]
server = "server.app:main"